requirements.txt
einops==0.6.0
fastai==2.7.10
hydra-core==1.3.1
imageio==2.22.4
matplotlib==3.5.3
numpy==1.21.6
omegaconf==2.2.3
opencv_python==4.6.0.66
openslide_python==1.2.0
pandas==1.3.5
scikit_image==0.19.3
scipy==1.7.3
seaborn==0.12.2
Shapely==1.8.5.post1
skimage==0.0
torch==1.13.0
torchmetrics==0.10.3
torchvision==0.14.0
tqdm==4.64.1
